+++ src/main/java/com/jianjin/assistant/service/graph/Entity.java
@@ -8,7 +8,7 @@ public class Entity {
     @JsonProperty("doc_hash")
     private String docHash;
     @JsonProperty("chunk_id")
-    private int chunkId;
+    private long chunkId;
 
     public Entity() {}
 
@@ -23,6 +23,6 @@ public class Entity {
     public void setType(EntityType type) { this.type = type; }
     public String getDocHash() { return docHash; }
     public void setDocHash(String docHash) { this.docHash = docHash; }
-    public int getChunkId() { return chunkId; }
-    public void setChunkId(int chunkId) { this.chunkId = chunkId; }
+    public long getChunkId() { return chunkId; }
+    public void setChunkId(long chunkId) { this.chunkId = chunkId; }
 }
